refactor(auth): AuthTokenManager를 인터페이스로 정의
- public 인터페이스가 명확하게 드러나도록 인터페이스로 역할을 정의
